+++ scripts/entry-points.mjs
@@ -10,6 +10,7 @@ export const entryPoints = {
   'carousel':            'src/stories/atoms/Carousel/Carousel.tsx',
   'checkbox':            'src/stories/atoms/Checkbox/Checkbox.tsx',
   'container':           'src/stories/atoms/Container/Container.tsx',
+  'columns':             'src/stories/atoms/Columns/Columns.tsx',
   'radio':               'src/stories/atoms/Radio/Radio.tsx',
   'icon':                'src/stories/atoms/Icon/Icon.tsx',
   'menu-button':         'src/stories/atoms/MenuButton/MenuButton.tsx',