+++ src/tokens/components/columns.ts
@@ -1,0 +1,4 @@
+import columnsJson from '../../../tokens/component/columns.json';
+import { flattenTokens } from '../utils';
+
+export const columnsTokens = flattenTokens(columnsJson as never);